refactor(cli): move Context plan choices into core
